scripts/requirements.txt
pymupdf>=1.24.0
python-docx>=1.1.0
openai>=1.50.0
supabase>=2.7.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
